% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{biggest_market_share}
\alias{biggest_market_share}
\title{Now, we can compute the airline market shares at each location}
\usage{
biggest_market_share(df, carrier_var, location_var)
}
\arguments{
\item{df}{data with flight records}

\item{carrier_var}{variable denoting the airline}

\item{location_var}{variable denoting the airport or city}
}
\value{
list of market shares
}
\description{
Now, we can compute the airline market shares at each location
}
