% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{busiest_routes}
\alias{busiest_routes}
\title{function for most popular air route (bidirectional)}
\usage{
busiest_routes(df, origin_var, dest_var)
}
\arguments{
\item{df}{data with flight records}

\item{origin_var}{variable denoting the origin airport or city}

\item{dest_var}{variable denoting the destination airport or city}
}
\value{
sorted list of busist routes
}
\description{
function for most popular air route (bidirectional)
}
