% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{function to load in air traffic data}
\usage{
load_data(air_data_path, carrier_data_path, airport_data_path)
}
\arguments{
\item{air_data_path}{data to load}

\item{carrier_data_path}{city ID to name}

\item{airport_data_path}{airport ID to name}
}
\description{
function to load in air traffic data
}
